# api/docs/swagger.yaml
definitions:
...
  rest.Languages:
    properties:
      en:
        type: string
      ru:
        type: string
      uz:
        type: string
    type: object
...
  rest.offerParams:
    properties:
      code:
        type: string
      name:
        $ref: '#/definitions/rest.Languages'
      should_Be_Signed:
        type: boolean
      signature_type:
        type: string
    required:
    - code
    - name
    type: object
...
  rest.regionParams:
    properties:
      code:
        type: string
      is_active:
        type: boolean
      rank:
        type: string
      title:
        $ref: '#/definitions/rest.Languages'
    required:
    - code
    - is_active
    - rank
    - title
    type: object
...
  rest.updateOfferParams:
    properties:
      name:
        $ref: '#/definitions/rest.Languages'
      should_Be_Signed:
        type: boolean
      signature_type:
        type: string
    required:
    - name
    type: object
...
  rest.updateRegionParams:
    properties:
      is_active:
        type: boolean
      rank:
        type: string
      title:
        $ref: '#/definitions/rest.Languages'
    required:
    - is_active
    - rank
    - title
    type: object
...
  views.Languages:
    properties:
      en:
        type: string
      ru:
        type: string
      uz:
        type: string
    type: object
...
  views.R:
    properties:
      data: {}
      error_code:
        type: integer
      error_note:
        type: string
      status:
        type: string
    type: object
...
  views.offer:
    properties:
      code:
        type: string
      name:
        $ref: '#/definitions/views.Languages'
      should_be_signed:
        type: boolean
      signature_type:
        type: string
    type: object
...
  views.region:
    properties:
      code:
        type: string
      is_active:
        type: boolean
      rank:
        type: string
      title:
        $ref: '#/definitions/views.Languages'
    type: object
info:
  contact:
    email: sodikjon.akhmedoff@gmail.com
    name: API Support
    url: https://github.com/s-akhmedoff
  description: CRUD API for all references
  license:
    name: Hamkorbank
    url: https://hamkorbank.uz/
  title: Reference Service v1
paths:
...
  /offers:
    get:
      consumes:
      - application/json
      description: API to get all offers
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/views.R'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/views.offer'
                  type: array
              type: object
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/views.R'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/views.R'
      summary: Read all offer
      tags:
      - Offer
    post:
      consumes:
      - application/json
      description: API to create a offer
      parameters:
      - description: create offer request parameters
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/rest.offerParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/views.R'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/views.R'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/views.R'
      summary: Create offer
      tags:
      - Offer
  /offers/{code}:
    delete:
      consumes:
      - application/json
      description: API to deactivate offer by code
      parameters:
      - description: offer code
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/views.R'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/views.R'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/views.R'
      summary: Delete offer
      tags:
      - Offer
    get:
      consumes:
      - application/json
      description: API to get a offer by code
      parameters:
      - description: offer code
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/views.R'
            - properties:
                data:
                  $ref: '#/definitions/views.offer'
              type: object
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/views.R'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/views.R'
      summary: Read one offer
      tags:
      - Offer
    put:
      consumes:
      - application/json
      description: API to update offer by code
      parameters:
      - description: offer code
        in: path
        name: code
        required: true
        type: string
      - description: update offer request parameters
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/rest.updateOfferParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/views.R'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/views.R'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/views.R'
      summary: Update offer
      tags:
      - Offer
...
  /regions:
    get:
      consumes:
      - application/json
      description: API to get all region
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/views.R'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/views.region'
                  type: array
              type: object
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/views.R'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/views.R'
      summary: Read all region
      tags:
      - Region
    post:
      consumes:
      - application/json
      description: API to create a region
      parameters:
      - description: create region request parameters
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/rest.regionParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/views.R'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/views.R'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/views.R'
      summary: Create region
      tags:
      - Region
  /regions/{code}:
    delete:
      consumes:
      - application/json
      description: API to deactivate region by code
      parameters:
      - description: region code
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/views.R'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/views.R'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/views.R'
      summary: Delete region
      tags:
      - Region
    get:
      consumes:
      - application/json
      description: API to get a region by code
      parameters:
      - description: region code
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/views.R'
            - properties:
                data:
                  $ref: '#/definitions/views.region'
              type: object
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/views.R'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/views.R'
      summary: Read one region
      tags:
      - Region
    put:
      consumes:
      - application/json
      description: API to update region by code
      parameters:
      - description: region code
        in: path
        name: code
        required: true
        type: string
      - description: update region request parameters
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/rest.updateRegionParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/views.R'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/views.R'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/views.R'
      summary: Update region
      tags:
      - Region
  /regions/import:
    put:
      consumes:
      - multipart/form-data
      description: API to deactivate region by code
      parameters:
      - description: 'Excel File in format: A:Code|B:NameRU|C:NameEn|D:NameUz|E:Rank'
        in: formData
        name: excel
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/views.R'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/views.R'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/views.R'
      summary: Create region
      tags:
      - Region
schemes:
- http
swagger: "2.0"